requirements.txt
streamlit
pyopenssl
cryptography
dnspython
